% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbwp_main.R
\name{bbwp}
\alias{bbwp}
\title{Calculate the BBWP scores on field and farm level}
\usage{
bbwp(
  B_SOILTYPE_AGR,
  B_LU_BRP,
  B_LU_BBWP,
  B_GWL_CLASS,
  B_SC_WENR,
  B_HELP_WENR,
  B_SLOPE,
  A_CLAY_MI,
  A_SAND_MI,
  A_SILT_MI,
  A_SOM_LOI,
  A_N_RT,
  A_FE_OX,
  A_AL_OX,
  A_P_CC,
  A_P_AL,
  A_P_WA,
  A_P_SG,
  B_GWP,
  B_AREA_DROUGHT,
  B_CT_PSW,
  B_CT_NSW,
  B_CT_PSW_MAX = 0.5,
  B_CT_NSW_MAX = 5,
  D_WP,
  D_RO_R,
  D_AREA,
  M_DRAIN,
  LSW,
  measures,
  sector,
  output = "all"
)
}
\arguments{
\item{B_SOILTYPE_AGR}{(character) The type of soil}

\item{B_LU_BRP}{(numeric) The crop type (conform BRP coding, preferable the most frequent crop on the field)}

\item{B_LU_BBWP}{(numeric) The BBWP category used for allocation of measures to BBWP crop categories}

\item{B_GWL_CLASS}{(character) The groundwater table class}

\item{B_SC_WENR}{(character) The risk for subsoil compaction as derived from risk assessment study of Van den Akker (2006)}

\item{B_HELP_WENR}{(character) The soil type abbreviation, derived from 1:50.000 soil map}

\item{B_SLOPE}{(numeric) The slope of the field (degrees)}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%)}

\item{A_SAND_MI}{(numeric) The sand content of the soil (\%)}

\item{A_SILT_MI}{(numeric) The silt content of the soil (\%)}

\item{A_SOM_LOI}{(numeric) The organic matter content of the soil (\%)}

\item{A_N_RT}{(numeric) The organic nitrogen content of the soil (mg N / kg)}

\item{A_FE_OX}{(numeric) The aluminium content of soil (mmol+ / kg)}

\item{A_AL_OX}{(numeric) The iron content of soil (mmol+ / kg)}

\item{A_P_CC}{(numeric) The plant available P content, measured via 0.01M CaCl2 extraction (mg / kg)}

\item{A_P_AL}{(numeric) The plant extractable P content, measured via ammonium lactate extraction (mg / kg)}

\item{A_P_WA}{(numeric) The P-content of the soil extracted with water}

\item{A_P_SG}{(numeric) The P-saturation index (\%)}

\item{B_GWP}{(boolean) is the field located in a groundwater protected area (options: TRUE or FALSE)}

\item{B_AREA_DROUGHT}{(boolean) is the field located in an area with high risks for water deficiencies (options: TRUE or FALSE)}

\item{B_CT_PSW}{(numeric) the critical target for required reduction in P loss from agriculture (kg P / ha) to reach targets of KRW}

\item{B_CT_NSW}{(numeric) the critical target for required reduction in N loss from agriculture (kg N / ha) to reach targets of KRW}

\item{B_CT_PSW_MAX}{(numeric) the max critical target for P reduction loss (kg P / ha)}

\item{B_CT_NSW_MAX}{(numeric) the max critical target for N reduction loss (kg N / ha)}

\item{D_WP}{(numeric) The fraction of the parcel that is surrounded by surface water}

\item{D_RO_R}{(numeric) The risk that surface water runs off the parcel}

\item{D_AREA}{(numeric) the area of the field (\ m2 or \ ha)}

\item{M_DRAIN}{(boolean) is there tube drainage present in the field}

\item{LSW}{(data.table) The surface water polygon for catchment or polder}

\item{measures}{(data.table) the measures planned / done per fields}

\item{sector}{(string) a vector with the farm type given the agricultural sector (options: 'melkveehouderij','akkerbouw','vollegrondsgroente','boomteelt','bollen','veehouderij','overig')}

\item{output}{(string) a vector specifying the output type of the function. Options: scores, measures}
}
\description{
Estimate the potential to contribute to agronomic and environmental challenges in a region for a farm and assess the impact of farm measures taken.
A high BBWP score is indicative for the number of opportunities to improve soil quality and land use.
}
