% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbwp_helpers.R
\name{bbwp_check_gt}
\alias{bbwp_check_gt}
\title{Helper function to check and update the groundwater class}
\usage{
bbwp_check_gt(B_GWL_CLASS, B_AER_CBS)
}
\arguments{
\item{B_GWL_CLASS}{(character) The groundwater table class}

\item{B_AER_CBS}{(character) The agricultural economic region in the Netherlands (CBS, 2016)}
}
\description{
Helper function to check and update the groundwater class
}
