% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbwp_helpers.R
\name{bbwp_check_lsw}
\alias{bbwp_check_lsw}
\title{Helper function to check the LSW table}
\usage{
bbwp_check_lsw(LSW, lat, lon)
}
\arguments{
\item{LSW}{(data.table) The input LSW table}

\item{lat}{(numeric) Latitude of the field (required if no LSW is submitted)}

\item{lon}{(numeric) Longitude of the field (required if no LSW is submitted)}
}
\description{
If there is no input LSW, the lon/lat are used to select values of the internal lws table
}
