% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbwp_helpers.R
\name{bbwp_check_meas}
\alias{bbwp_check_meas}
\title{Helper function to check, update and extend the table with measurements}
\usage{
bbwp_check_meas(dt, eco = FALSE, score = TRUE)
}
\arguments{
\item{dt}{(data.table) a data.table containing the bbwp measures with or without properties}

\item{eco}{(boolean) get only the measures for Ecoregeling (with a given eco_id)}

\item{score}{(boolean) is check done for scoring of subsets of full list of measures}
}
\description{
Helper function to check, update and extend the table with measurements
}
