% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbwp_field_scores.R
\name{bbwp_field_scores}
\alias{bbwp_field_scores}
\title{Calculate the total score of five opportunity indicators for all fields in the Netherlands}
\usage{
bbwp_field_scores(
  B_SOILTYPE_AGR,
  B_GWL_CLASS,
  A_P_SG,
  B_SLOPE_DEGREE,
  B_LU_BBWP,
  B_AER_CBS,
  M_DRAIN,
  D_SA_W,
  D_RISK_NGW,
  D_RISK_NSW,
  D_RISK_PSW,
  D_RISK_NUE,
  D_RISK_WB,
  B_GWP,
  B_AREA_DROUGHT,
  B_CT_PSW,
  B_CT_NSW,
  B_CT_PSW_MAX = 0.5,
  B_CT_NSW_MAX = 5,
  measures,
  sector
)
}
\arguments{
\item{B_SOILTYPE_AGR}{(character) The type of soil}

\item{B_GWL_CLASS}{(character) The groundwater table class}

\item{A_P_SG}{(numeric)}

\item{B_SLOPE_DEGREE}{(numeric) The slope of the field (degrees)}

\item{B_LU_BBWP}{(character) The BBWP category used for allocation of measures to BBWP crop categories}

\item{B_AER_CBS}{(character) The agricultural economic region in the Netherlands (CBS, 2016)}

\item{M_DRAIN}{(boolean) is there tube drainage present in the field}

\item{D_SA_W}{(numeric) The wet perimeter index of the field, fraction that field is surrounded by water}

\item{D_RISK_NGW}{(numeric) the risk for nitrate leaching to groundwater given field properties}

\item{D_RISK_NSW}{(numeric) the risk for nitrate leaching and runoff to surface water given field properties}

\item{D_RISK_PSW}{(numeric) the risk for phosphorus leaching and runoff to surface water given field properties}

\item{D_RISK_NUE}{(numeric) the efficiency of nitrogen and phosphorus fertilizer use given field properties}

\item{D_RISK_WB}{(numeric) the potential to buffer and store water and efficiently use water for plant growth given field properties}

\item{B_GWP}{(boolean) is the field located in a groundwater protected area (options: TRUE or FALSE)}

\item{B_AREA_DROUGHT}{(boolean) is the field located in an area with high risks for water deficiencies (options: TRUE or FALSE)}

\item{B_CT_PSW}{(numeric) the critical target for required reduction in P loss from agriculture (kg P / ha) to reach targets of KRW}

\item{B_CT_NSW}{(numeric) the critical target for required reduction in N loss from agriculture (kg N / ha) to reach targets of KRW}

\item{B_CT_PSW_MAX}{(numeric) the max critical target for P reduction loss (kg P / ha)}

\item{B_CT_NSW_MAX}{(numeric) the max critical target for N reduction loss (kg N / ha)}

\item{measures}{(data.table) the measures planned / done per fields}

\item{sector}{(string) a vector with the farm type given the agricultural sector (options: 'dairy', 'arable', 'tree_nursery', 'bulbs')}
}
\description{
Estimate the potential to contribute to agronomic and environmental challenges in a region
}
