% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbwp_helpers.R
\name{bbwp_format_aer}
\alias{bbwp_format_aer}
\title{Convert possible B_AER_CBS values to standardized values}
\usage{
bbwp_format_aer(B_AER_CBS)
}
\arguments{
\item{B_AER_CBS}{(character) The agricultural economic region in the Netherlands (CBS, 2016)}
}
\value{
A standardized B_AER_CBS value as required for the BBWP functions. A character string.
}
\description{
This function formats information of Agricultural Economic Region so it can be understood by other functions
}
\examples{
bbwp_format_aer(c("LG13","LG12"))
bbwp_format_aer(c("LG13","LG12",'Rivierengebied'))

}
