% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbwp_helpers.R
\name{bbwp_format_sc_wenr}
\alias{bbwp_format_sc_wenr}
\title{Convert possible B_SC_WENR values to standardized values}
\usage{
bbwp_format_sc_wenr(B_SC_WENR)
}
\arguments{
\item{B_SC_WENR}{(character) The risk for subsoil compaction as derived from risk assessment study of Van den Akker (2006)}
}
\value{
A standardized B_SC_WENR value as required for the BBWP functions. A character string.
}
\description{
This function formats information of subsoil compaction so it can be understood by BBWP functions
}
\examples{
bbwp_format_sc_wenr(c("1","5"))
bbwp_format_sc_wenr(c("Zeer beperkt","Water",'4'))

}
