% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbwp_meas_rank.R
\name{bbwp_meas_rank}
\alias{bbwp_meas_rank}
\title{Rank the suitability of agronomic measures to improve soil and water management for a given field}
\usage{
bbwp_meas_rank(
  B_SOILTYPE_AGR,
  B_GWL_CLASS,
  A_P_SG,
  B_SLOPE_DEGREE,
  B_LU_BBWP,
  B_AER_CBS,
  M_DRAIN,
  D_SA_W,
  S_BBWP_NGW,
  S_BBWP_NSW,
  S_BBWP_PSW,
  S_BBWP_NUE,
  S_BBWP_WB,
  measures,
  sector
)
}
\arguments{
\item{B_SOILTYPE_AGR}{(character) The type of soil}

\item{B_GWL_CLASS}{(character) The groundwater table class}

\item{A_P_SG}{(numeric)}

\item{B_SLOPE_DEGREE}{(numeric) The slope of the field (degrees)}

\item{B_LU_BBWP}{(character) The BBWP category used for allocation of measures to BBWP crop categories}

\item{B_AER_CBS}{(character) The agricultural economic region in the Netherlands (CBS, 2016)}

\item{M_DRAIN}{(boolean) is there tube drainage present in the field}

\item{D_SA_W}{(numeric) The wet perimeter index of the field, fraction that field is surrounded by water}

\item{S_BBWP_NGW}{(numeric) the BBWP score for nitrate leaching to groundwater given field properties (0-100, with 100 equals targets met)}

\item{S_BBWP_NSW}{(numeric) the BBWP score for nitrate leaching and runoff to surface water given field properties (0-100, with 100 equals targets met)}

\item{S_BBWP_PSW}{(numeric) the BBWP score for phosphorus leaching and runoff to surface water given field properties (0-100, with 100 equals targets met)}

\item{S_BBWP_NUE}{(numeric) the BBWP score for potential improvement of the efficiency of nitrogen and phosphorus fertilizer use given field properties (0-100, with 100 equals targets met)}

\item{S_BBWP_WB}{(numeric) the BBWP score for potential improvement of the potential to buffer and store water and efficiently use water for plant growth given field properties (0-100, with 100 equals targets met)}

\item{measures}{(data.table) table with the properties of the available measures}

\item{sector}{(string) a vector with the farm type given the agricultural sector (options: 'diary', 'arable', 'tree_nursery', 'bulbs')}
}
\description{
Estimate the score for agronomic measures to improve soil and water management on agricultural farms. 
And send an ordered list back of the most suitable measures.
}
