% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbwp_meas_scores.R
\name{bbwp_meas_score}
\alias{bbwp_meas_score}
\title{Evaluate the contribution of agronomic measures to improve soil mand water management}
\usage{
bbwp_meas_score(
  B_SOILTYPE_AGR,
  B_GWL_CLASS,
  A_P_CC,
  A_P_AL,
  B_SLOPE_DEGREE,
  B_LU_BBWP,
  B_AER_CBS,
  M_DRAIN,
  D_SA_W,
  D_OPI_NGW,
  D_OPI_NSW,
  D_OPI_PSW,
  D_OPI_NUE,
  D_OPI_WB,
  measures = NULL,
  sector,
  B_LS_HYDROCAT
)
}
\arguments{
\item{B_SOILTYPE_AGR}{(character) The type of soil}

\item{B_GWL_CLASS}{(character) The groundwater table class}

\item{A_P_CC}{(numeric) The plant available P content, measured via 0.01M CaCl2 extraction (mg / kg)}

\item{A_P_AL}{(numeric) The plant extractable P content, measured via ammonium lactate extraction (mg / kg)}

\item{B_SLOPE_DEGREE}{(numeric) The slope of the field (degrees)}

\item{B_LU_BBWP}{(numeric) The BBWP category used for allocation of measures to BBWP crop categories}

\item{B_AER_CBS}{(character) The agricultural economic region in the Netherlands (CBS, 2016)}

\item{M_DRAIN}{(boolean) is there tube drainage present in the field}

\item{D_SA_W}{(numeric) The wet perimeter index of the field, fraction that field is surrounded by water}

\item{D_OPI_NGW}{(numeric) the opportunity index (risk x impact) for nitrate leaching to groundwater given field properties}

\item{D_OPI_NSW}{(numeric) the opportunity index (risk x impact) for nitrate leaching and runoff to surface water given field properties}

\item{D_OPI_PSW}{(numeric) the opportunity index (risk x impact) for phosphorus leaching and runoff to surface water given field properties}

\item{D_OPI_NUE}{(numeric) the opportunity index (risk x impact) to improve the efficiency of nitrogen and phosphorus fertilizer use given field properties}

\item{D_OPI_WB}{(numeric) the opportunity index (risk x impact) to improve the potential to buffer and store water and efficiently use water for plant growth given field properties}

\item{measures}{(data.table) the measures planned / done per fields}

\item{sector}{(string) a vector with the farm type given the agricultural sector (options: 'dairy', 'arable', 'tree_nursery', 'bulbs')}

\item{B_LS_HYDROCAT}{(character) Landscape category for differentiating effect of measures on water buffering.
(options: "hoge_gronden", "flanken", "beekdalen", "lokale_laagtes", "polders")}
}
\description{
Estimate the score for agronomic measures taken to improve soil and water management on agricultural farms.
}
