% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbwp_tables.R
\docType{data}
\name{bbwp_parms}
\alias{bbwp_parms}
\title{A table with the BBWP parameters}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 26 rows and 14 columns.
}
\usage{
bbwp_parms
}
\description{
This table contains the BBWP variables and their possible values according to pandex to standardize checks
}
\details{
\describe{
 \item{code}{The parameter code}
 \item{parameter}{Brief description of the parameter}
 \item{unit}{The unit of the parameter if applicable}
 \item{product}{Data classifier, A = Soil measurements, B = Environmental characteristics,
  D = Soil or feed characteristics derived from soil/feed measurements,
  M = Soil management measures, S = Scores, RM = Recommendations (measures/gifts),
  I = Indicators, F = Feed measurements, P = soil amendment product}
 \item{element}{Indicates the chemical element or parameter name}
 \item{method1}{Method used to determine value}
 \item{method2}{Additional details on method}
 \item{data_type}{Type of data the parameter pertains to: numeric, integer, char, bool, geom, enum}
 \item{value_min}{Lowest possible value the parameter may have if numeric or integer}
 \item{value_max}{Highest possible value the parameter may have if numeric or integer}
 \item{explanation}{Some additional explanation}
 \item{enum}{boolean whether parameter values are drawn from a limited set}
 \item{options}{Allowed values for a parameter of type enum seperated by "||"}
 \item{choices}{Vectorized version of options}
 }
}
\keyword{datasets}
