% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBWPC-package.R
\docType{package}
\name{BBWPC-package}
\alias{BBWPC}
\alias{BBWPC-package}
\title{BBWPC: Calculator for BedrijfsBodemWaterPlan (BBWP)}
\description{
The BBWP Calculator can be used to calculate the opportunities and suitability of agronomic measures to improve the sustainability of agriculture. It includes the framework of the BedrijfsBodemWaterPlan (BBWP) developed in the Netherlands (Ros et al., 2020), as well the Dutch Ecoregelingen Framework developed by Boerennatuur and associated agricultural farmers collectives. Both come together in the BBWP tool to support farmers in their soil and water management. The BedrijfsBodemWaterPlan is a tool for The Netherlands that recommends field specific agronomic measures focusing on the improvement of soil health, the reduction of nutrient losses to groundwater and surface water and the water retention properties. More information about the BedrijfsBodemWaterPlan can be found at www.bedrijfsbodemwaterplan.nl.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nmi-agro/BedrijfsBodemWaterPlanCalculator}
  \item \url{https://nmi-agro.github.io/BedrijfsBodemWaterPlanCalculator/}
  \item Report bugs at \url{https://github.com/nmi-agro/BedrijfsBodemWaterPlanCalculator/issues}
}

}
\author{
\strong{Maintainer}: Gerard Ros \email{gerard.ros@nmi-agro.nl}

Authors:
\itemize{
  \item Gerard Ros \email{gerard.ros@nmi-agro.nl}
  \item Sven Verweij \email{sven.verweij@nmi-agro.nl}
}

Other contributors:
\itemize{
  \item Job de Pater \email{job.depater@nmi-agro.nl} [contributor]
  \item Yuki Fujita \email{yuki.fujita@nmi-agro.nl} [contributor]
  \item Kees van den Dool \email{kees.vandendool@nmi-agro.nl} [contributor]
  \item Astrid Berndsen \email{astrid.berndsen@nmi-agro.nl} [contributor]
  \item Brent Riechelman \email{brent.riechelman@nmi-agro.nl} (\href{https://orcid.org/0000-0003-1501-1573}{ORCID}) [contributor]
  \item Nutriënten Managment Instituut \email{nmi@nmi-agro.nl} [copyright holder]
}

}
\keyword{internal}
