% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbwp_helpers.R
\name{cdf_rank}
\alias{cdf_rank}
\title{Calculate the chance that soil property is lower than the measured value given the variation in a catchment}
\usage{
cdf_rank(smean, ssd, svalue)
}
\arguments{
\item{smean}{(numeric) The mean value for a numeric soil property in a given catchment}

\item{ssd}{(numeric) The variation (standard deviation) among a numeric soil property in a given catchment}

\item{svalue}{(numeric) The measured value of a soil property of a field inside a catchment}
}
\description{
assuming a normal distributed density function. The chance can be used as a relative ranking position of the field given the variation inside the catchment.
}
