% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_farm_score.R
\name{er_farm_score}
\alias{er_farm_score}
\title{Calculate the total EcoRegelingen score of five opportunity indicators}
\usage{
er_farm_score(
  D_OPI_TOT,
  D_OPI_SOIL,
  D_OPI_WATER,
  D_OPI_CLIMATE,
  D_OPI_BIO,
  D_OPI_LANDSCAPE,
  D_AREA
)
}
\arguments{
\item{D_OPI_TOT}{(numeric) the Ecoregeling score for the integrative opportunity index for each field}

\item{D_OPI_SOIL}{(numeric) the Ecoregeling scoring index for soil quality for each field}

\item{D_OPI_WATER}{(numeric) the Ecoregeling scoring index for water quality for each field}

\item{D_OPI_CLIMATE}{(numeric) the Ecoregeling scoring index for climate for each field}

\item{D_OPI_BIO}{(numeric) the Ecoregeling scoring index for biodiversity for each field}

\item{D_OPI_LANDSCAPE}{(numeric) the Ecoregeling scoring index for landscape for each field}

\item{D_AREA}{(numeric) the area of the field (\ m2 or \ ha)}
}
\description{
Estimate the potential to contribute to agronomic and environmental challenges in a region for a farm
}
