% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbwp_tables.R
\docType{data}
\name{er_measures}
\alias{er_measures}
\title{The applicability of Ecoregeling Measures per crop type}
\format{
A data.table with 6731 rows and 3 columns:
\describe{
  \item{eco_id}{ecoregeling measure id}
  \item{B_lU_BRP}{dutch crop id from the BRP}
  \item{eco_app}{a helper variable telling that the measure is applicable for that crop}
}
}
\usage{
er_measures
}
\description{
This table contains the crops for which the Ecoregeling method counts scores and rewards
}
\keyword{datasets}
