% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_opi.R
\name{er_opi}
\alias{er_opi}
\title{Calculate the Opportunity Index for the Ecoregeling method}
\usage{
er_opi(
  B_SOILTYPE_AGR,
  S_ER_SOIL,
  S_ER_WATER,
  S_ER_CLIMATE,
  S_ER_BIODIVERSITY,
  S_ER_LANDSCAPE,
  S_ER_REWARD,
  B_AREA,
  medalscore
)
}
\arguments{
\item{B_SOILTYPE_AGR}{(character) The type of soil}

\item{S_ER_SOIL}{(numeric) the Ecoregeling scoring index for soil quality for each field}

\item{S_ER_WATER}{(numeric) the Ecoregeling scoring index for water quality for each field}

\item{S_ER_CLIMATE}{(numeric) the Ecoregeling scoring index for climate for each field}

\item{S_ER_BIODIVERSITY}{(numeric) the Ecoregeling scoring index for biodiversity for each field}

\item{S_ER_LANDSCAPE}{(numeric) the Ecoregeling scoring index for landscape for each field}

\item{S_ER_REWARD}{(numeric) The financial reward per field for taking Ecoregeling measures (euro / ha)}

\item{B_AREA}{(numeric) the area of the field (m2)}

\item{medalscore}{(character) The desired medal score expressed as bronze, silver or gold}
}
\description{
Estimate the contribution of individual fields to the desired scores for highest impact on farm level as well as the total farm score, weighted given the distance to target.
A high Ecoregeling score is indicative for the number of opportunities to improve soil quality, water quality, climate biodiversity and landscape.
}
