% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_pdf.R
\name{er_pdf}
\alias{er_pdf}
\title{Make table of measures that generated scores}
\usage{
er_pdf(
  croprotation,
  measurescores,
  dt.field.measures,
  dt.farm.measures,
  B_AREA
)
}
\arguments{
\item{croprotation}{(boolean) is the pdf requested in function er_croprotation}

\item{measurescores}{(boolean) is the pdf requested in function er_meas_scores}

\item{dt.field.measures}{(datatable) field measures}

\item{dt.farm.measures}{(datatable) farm measures}

\item{B_AREA}{(numeric) the area of the field (m2)}
}
\description{
Present each measure that generated scores on field and/or farm level in table format with scores for soil quality, water quality, climate, biodiversity and landscape given soil type.
}
