% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbwp_tables.R
\docType{data}
\name{er_scoring}
\alias{er_scoring}
\title{The importance and scoring of environmental challenges for Ecoregelingen}
\format{
A data.table with 8 rows and 7 columns:
\describe{
  \item{soiltype}{bla bla}
  \item{type}{bla bla}
  \item{cf_soil}{bla bla}
  \item{cf_water}{bla bla}
  \item{cf_climate}{bla bla}
  \item{cf_biodiversity}{bla bla}
  \item{cf_landscape}{ bla bla}
}
}
\usage{
er_scoring
}
\description{
This table contains the objective scoring and correction factors to estimate score for Ecoregeling
}
\keyword{datasets}
