% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbwp_tables.R
\docType{data}
\name{lsw}
\alias{lsw}
\title{A national dataset describing statistics of soil properties per Local Surface Water}
\format{
A data.table with 546 rows and 32 columns:
\describe{
  \item{c1}{bla bla}
  \item{c2}{bla bla}
  \item{c3}{bla bla}
  \item{c4}{bla bla}
}
}
\usage{
lsw
}
\description{
This table contains mean and standard deviation of soil properties per Local Surface Water Polygon in the Netherlands
}
\keyword{datasets}
