% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbwp_helpers.R
\name{wf}
\alias{wf}
\title{Helper function to weight and correct the risk and scores}
\usage{
wf(x, type = "indicators", penalty = TRUE)
}
\arguments{
\item{x}{The risk or score value to be weighted}

\item{type}{Use the weighing function for indicators or score}

\item{penalty}{(boolean) the option to apply a penalty for high risk BBWP field indicators}
}
\description{
Helper function to weight and correct the risk and scores
}
