% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbwp_calc_psp.R
\name{bbwp_calc_psp}
\alias{bbwp_calc_psp}
\title{Calculate precipitation surplus}
\usage{
bbwp_calc_psp(B_LU_BRP, M_GREEN, M_GREEN_START = 10L, M_GREEN_TERMINATE = 3L)
}
\arguments{
\item{B_LU_BRP}{(numeric) The crop code from the BRP}

\item{M_GREEN}{(boolean) A soil measure. Are catch crops sown after main crop (optional, options: TRUE, FALSE)}

\item{M_GREEN_START}{(integer) Month in which the green manure is sown}

\item{M_GREEN_TERMINATE}{(integer) Month in which the green manure is terminated, this month is included as month with green manure}

\item{...}{
  Arguments passed on to \code{\link[OBIC:calc_psp]{OBIC::calc_psp}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
Calculate precipitation surplus
}
\seealso{
\code{\link[OBIC]{calc_psp}}
}
