% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbwp_farm_score.R
\name{bbwp_farm_score}
\alias{bbwp_farm_score}
\title{Calculate the total score of five opportunity indicators for all fields in the Netherlands}
\usage{
bbwp_farm_score(
  D_OPI_TOT,
  D_OPI_NGW,
  D_OPI_NSW,
  D_OPI_PSW,
  D_OPI_NUE,
  D_OPI_WB,
  D_AREA
)
}
\arguments{
\item{D_OPI_TOT}{(numeric) the score for the integrative opportunity index (risk times impact) for each field}

\item{D_OPI_NGW}{(numeric) the scoring index for lowering N emission to ground water (risk times impact) for each field}

\item{D_OPI_NSW}{(numeric) the scoring index for lowering N emission to surface water (risk times impact) for each field}

\item{D_OPI_PSW}{(numeric) the scoring index for lowering P emission to surface water (risk times impact) for each field}

\item{D_OPI_NUE}{(numeric) the scoring index to use N and P inputs efficiently for each field}

\item{D_OPI_WB}{(numeric) the scoring index to buffer and store water and efficiently use water for plant growth for each field}

\item{D_AREA}{(numeric) the area of the field (\ m2 or \ ha)}
}
\description{
Estimate the potential to contribute to agronomic and environmental challenges in a region for a farm
}
