% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbwp_field_indicators.R
\name{bbwp_field_indicators}
\alias{bbwp_field_indicators}
\title{Calculate five risk or opportunity indicators for an agricultural field}
\usage{
bbwp_field_indicators(
  D_NGW_SCR,
  D_NGW_LEA,
  D_NGW_NLV,
  D_NSW_SCR,
  D_NSW_GWT,
  D_NSW_RO,
  D_NSW_WS,
  D_NSW_NLV,
  D_PSW_SCR,
  D_PSW_GWT,
  D_PSW_RO,
  D_PSW_WS,
  D_PSW_PCC,
  D_PSW_PSG,
  D_PSW_PRET,
  D_NUE_WRI,
  D_NUE_PBI,
  D_NUE_WDRI,
  D_NUE_NLV,
  D_WUE_WWRI,
  D_WUE_WDRI,
  D_WUE_WHC
)
}
\arguments{
\item{D_NGW_SCR}{(numeric) The relative score of soil compaction risk for N loss to groundwater}

\item{D_NGW_LEA}{(numeric) The relative score of nitrate leaching risk for N loss to groundwater}

\item{D_NGW_NLV}{(numeric) The relative score of N mineralization for N loss to groundwater}

\item{D_NSW_SCR}{(numeric) The relative score of soil compaction risk for N loss to surface water}

\item{D_NSW_GWT}{(numeric) The relative score of soil wetness for N loss to surface water}

\item{D_NSW_RO}{(numeric) The relative score of runoff risks for N loss to surface water}

\item{D_NSW_WS}{(numeric) The relative score of wet surrounding for N loss to surface water}

\item{D_NSW_NLV}{(numeric) The relative score of N mineralization for N loss to surface water}

\item{D_PSW_SCR}{(numeric) The relative score of soil compaction risk for P loss to surface water}

\item{D_PSW_GWT}{(numeric) The relative score of soil wetness for P loss to surface water}

\item{D_PSW_RO}{(numeric) The relative score of runoff risks for P loss to surface water}

\item{D_PSW_WS}{(numeric) The relative score of wet surrounding for P loss to surface water}

\item{D_PSW_PCC}{(numeric) The relative score of P levels in soil solution for P loss to surface water}

\item{D_PSW_PSG}{(numeric) The relative score of P saturation degree for P loss to surface water}

\item{D_PSW_PRET}{(numeric) The relative score of P retention capacity for P loss to surface water}

\item{D_NUE_WRI}{(numeric) The relative score of soil water stress for improved eficiency of N and P inputs}

\item{D_NUE_PBI}{(numeric) The relative score of P availability for improved eficiency of N and P inputs}

\item{D_NUE_WDRI}{(numeric) The relative score of drought stress for improved eficiency of N and P inputs}

\item{D_NUE_NLV}{(numeric) The relative score of N mineralization for improved eficiency of N and P inputs}

\item{D_WUE_WWRI}{(numeric) The relative score of soil wetness stress for improved eficiency of water}

\item{D_WUE_WDRI}{(numeric) The relative score of drought stress for improved eficiency of water}

\item{D_WUE_WHC}{(numeric) The relative score of drought stress for improved eficiency of water}
}
\description{
These include indicators: 
for limited efficiency of nitrogen and phosphorus; a high risk is also indicative of high potential to improve NUE.
for the potential to buffer and store water and efficiently use water for plant growth.
for the risk for nitrogen and phosphorus leaching and runoff to surface water.
for the risk for nitrate leaching to groundwater.
}
