% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbwp_field_properties.R
\name{bbwp_field_properties}
\alias{bbwp_field_properties}
\title{Calculate the relative impact of field properties given its position in the landscape}
\usage{
bbwp_field_properties(
  B_SOILTYPE_AGR,
  B_LU_BRP,
  B_GWL_CLASS,
  B_SC_WENR,
  B_HELP_WENR,
  A_CLAY_MI,
  A_SAND_MI,
  A_SILT_MI,
  A_SOM_LOI,
  A_N_RT,
  A_FE_OX,
  A_AL_OX,
  A_P_CC,
  A_P_AL,
  A_P_WA,
  A_P_SG,
  D_WP,
  D_RO_R,
  LSW
)
}
\arguments{
\item{B_SOILTYPE_AGR}{(character) The type of soil}

\item{B_LU_BRP}{(numeric) The crop type (conform BRP coding, preferable the most frequent crop on the field)}

\item{B_GWL_CLASS}{(character) The groundwater table class}

\item{B_SC_WENR}{(character) The risk for subsoil compaction as derived from risk assessment study of Van den Akker (2006)}

\item{B_HELP_WENR}{(character) The soil type abbreviation, derived from 1:50.000 soil map}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%)}

\item{A_SAND_MI}{(numeric) The sand content of the soil (\%)}

\item{A_SILT_MI}{(numeric) The silt content of the soil (\%)}

\item{A_SOM_LOI}{(numeric) The organic matter content of the soil (\%)}

\item{A_N_RT}{(numeric) The organic nitrogen content of the soil (mg N / kg)}

\item{A_FE_OX}{(numeric) The aluminium content of soil (mmol+ / kg)}

\item{A_AL_OX}{(numeric) The iron content of soil (mmol+ / kg)}

\item{A_P_CC}{(numeric) The plant available P content, measured via 0.01M CaCl2 extraction (mg / kg)}

\item{A_P_AL}{(numeric) The plant extractable P content, measured via ammonium lactate extraction (mg / kg)}

\item{A_P_WA}{(numeric) The P-content of the soil extracted with water}

\item{A_P_SG}{(numeric) The P-saturation index (/%)}

\item{D_WP}{(numeric) The fraction of the parcel that is surrounded by surface water}

\item{D_RO_R}{(numeric) The risk that surface water runs off the parcel}

\item{LSW}{(data.table) The surface water polygon for catchment or polder}
}
\description{
Estimate the relative ranking of field properties given their contribution to nutrient losses to aquatic ecosystems as well as nutrient and water efficiency.
A high rank is indicative for the number of opportunities to improve soil quality and land use.
}
