% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbwp_meas_scores.R
\name{bbwp_meas_rank}
\alias{bbwp_meas_rank}
\title{Rank the suitability of agronomic measures to improve soil and water management for a given field}
\usage{
bbwp_meas_rank(
  B_SOILTYPE_AGR,
  B_GWL_CLASS,
  A_P_SG,
  B_SLOPE,
  B_LU_BRP,
  M_DRAIN,
  D_WP,
  D_OPI_NGW,
  D_OPI_NSW,
  D_OPI_PSW,
  D_OPI_NUE,
  D_OPI_WB,
  available_measures,
  sector
)
}
\arguments{
\item{B_SOILTYPE_AGR}{(character) The type of soil}

\item{B_GWL_CLASS}{(character) The groundwater table class}

\item{A_P_SG}{(numeric)}

\item{B_SLOPE}{(boolean)}

\item{B_LU_BRP}{(integer)}

\item{M_DRAIN}{(boolean) is there tube drainage present in the field}

\item{D_WP}{(numeric) The fraction of the parcel that is surrounded by surface water}

\item{D_OPI_NGW}{(numeric) the opportunity index (risk x impact) for nitrate leaching to groundwater given field properties}

\item{D_OPI_NSW}{(numeric) the opportunity index (risk x impact) for nitrate leaching and runoff to surface water given field properties}

\item{D_OPI_PSW}{(numeric) the opportunity index (risk x impact) for phosphorus leaching and runoff to surface water given field properties}

\item{D_OPI_NUE}{(numeric) the opportunity index (risk x impact) to improve the efficiency of nitrogen and phosphorus fertilizer use given field properties}

\item{D_OPI_WB}{(numeric) the opportunity index (risk x impact) to improve the potential to buffer and store water and efficiently use water for plant growth given field properties}

\item{available_measures}{(data.table) table with the properties of the available measures}

\item{sector}{(string) a vector with the farm type given the agricultural sector (options: 'diary', 'arable', 'tree_nursery', 'bulbs')}
}
\description{
Estimate the score for agronomic measures to improve soil and water management on agricultural farms. 
And send an ordered list back of the most suitable measures.
}
