% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBWPC-package.R
\docType{package}
\name{BBWPC-package}
\alias{BBWPC}
\alias{BBWPC-package}
\title{BBWPC: Calculator for BedrijfsBodemWaterPlan (BBWP)}
\description{
This package can be used to calculate the opportunities and suitability of agronomic measures (BBWP). The BedrijfsBodemWaterPlan is a tool for The Netherlands that recommends field specific agronomic measures focusing on the improvement of soil health. the reduction of nutrient losses to groundwater and surface water and the water retention properties. More information about the BedrijfsBodemWaterPlan can be found a bedijfsbodemwaterplan.nl.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AgroCares/Soil-Water-Plan-Calculator}
  \item Report bugs at \url{https://github.com/AgroCares/Soil-Water-Plan-Calculator/issues}
}

}
\author{
\strong{Maintainer}: Gerard Ros \email{gerard.ros@nmi-agro.nl}

Authors:
\itemize{
  \item Sven Verweij \email{sven.verweij@nmi-agro.nl}
}

Other contributors:
\itemize{
  \item Yuki Fujita \email{yuki.fujita@nmi-agro.nl} [contributor]
  \item Nutriënten Managment Instituut \email{nmi@nmi-agro.nl} [copyright holder]
}

}
\keyword{internal}
