% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_main.R
\name{ecoregeling}
\alias{ecoregeling}
\title{Calculate the Ecoregeling scores on field and farm level}
\usage{
ecoregeling(
  B_SOILTYPE_AGR,
  B_LU_BRP,
  B_LU_BBWP,
  B_GWL_CLASS,
  B_SLOPE,
  A_P_SG,
  D_WP,
  D_AREA,
  M_DRAIN,
  farmscore,
  measures,
  sector,
  output = "scores"
)
}
\arguments{
\item{B_SOILTYPE_AGR}{(character) The type of soil}

\item{B_LU_BRP}{(numeric) The crop type (conform BRP coding, preferable the most frequent crop on the field)}

\item{B_LU_BBWP}{(numeric) The BBWP category used for allocation of measures to BBWP crop categories}

\item{B_GWL_CLASS}{(character) The groundwater table class}

\item{B_SLOPE}{(numeric) The slope of the field (degrees)}

\item{A_P_SG}{(numeric) The P-saturation index (\%)}

\item{D_WP}{(numeric) The fraction of the parcel that is surrounded by surface water}

\item{D_AREA}{(numeric) the area of the field (\ m2 or \ ha)}

\item{M_DRAIN}{(boolean) is there tube drainage present in the field}

\item{farmscore}{(numeric) The desired total ER score on farm level}

\item{measures}{(data.table) the measures planned / done per fields}

\item{sector}{(string) a vector with the farm type given the agricultural sector (options: 'melkveehouderij','akkerbouw','vollegrondsgroente','boomteelt','bollen','veehouderij','overig')}

\item{output}{(string) a vector specifying the output type of the function. Options: scores, measures}
}
\description{
Estimate the potential to contribute to agronomic and environmental challenges in a region for a farm and assess the impact of farm measures taken.
A high Ecoregeling score is indicative for the number of opportunities to improve soil quality, water quality, climate biodiversity and landscape.
}
