% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_main.R
\name{ecoregeling}
\alias{ecoregeling}
\title{Calculate the Ecoregeling scores on field and farm level}
\usage{
ecoregeling(
  B_SOILTYPE_AGR,
  B_LU_BRP,
  B_LU_BBWP,
  B_GWL_CLASS,
  B_SLOPE_DEGREE,
  B_AER_CBS,
  A_P_SG,
  D_SA_W,
  B_AREA,
  M_DRAIN,
  farmscore,
  measures,
  sector,
  output = "scores"
)
}
\arguments{
\item{B_SOILTYPE_AGR}{(character) The type of soil}

\item{B_LU_BRP}{(numeric) The crop type (conform BRP coding, preferable the most frequent crop on the field)}

\item{B_LU_BBWP}{(numeric) The BBWP category used for allocation of measures to BBWP crop categories}

\item{B_GWL_CLASS}{(character) The groundwater table class}

\item{B_SLOPE_DEGREE}{(numeric) The slope of the field (degrees)}

\item{B_AER_CBS}{(character) The agricultural economic region in the Netherlands (CBS, 2016)}

\item{A_P_SG}{(numeric) The P-saturation index (\%)}

\item{D_SA_W}{(numeric) The wet perimeter index of the field, fraction that field is surrounded by water}

\item{B_AREA}{(numeric) the area of the field (m2)}

\item{M_DRAIN}{(boolean) is there tube drainage present in the field}

\item{farmscore}{(numeric) The desired total ER score on farm level}

\item{measures}{(data.table) the measures planned / done per fields}

\item{sector}{(string) a vector with the farm type given the agricultural sector (options: options: 'diary', 'arable', 'tree_nursery', 'bulbs')}

\item{output}{(string) a vector specifying the output type of the function. Options: scores, measures}
}
\description{
Estimate the potential to contribute to agronomic and environmental challenges in a region for a farm and assess the impact of farm measures taken.
A high Ecoregeling score is indicative for the number of opportunities to improve soil quality, water quality, climate biodiversity and landscape.
}
