% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbwp_tables.R
\docType{data}
\name{er_aer_reward}
\alias{er_aer_reward}
\title{An overview of correction factors per economic agricultural area}
\format{
A data.table with 14 rows and 3 columns:
\describe{
  \item{statcode}{bla bla}
  \item{statname}{bla bla}
  \item{er_cf}{bla bla}
}
}
\usage{
er_aer_reward
}
\description{
This table contains the correction factor for the financial reward differentiated per Agricultural Economic Region
}
\keyword{datasets}
