% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_croprotation.R
\name{er_croprotation}
\alias{er_croprotation}
\title{Calculate the crop rotation based total score for five opportunity indicators}
\usage{
er_croprotation(
  B_SOILTYPE_AGR,
  B_LU_BRP,
  B_LU_BBWP,
  B_AER_CBS,
  D_AREA,
  B_CT_SOIL,
  B_CT_WATER,
  B_CT_CLIMATE,
  B_CT_BIO,
  B_CT_LANDSCAPE
)
}
\arguments{
\item{B_SOILTYPE_AGR}{(character) The type of soil}

\item{B_LU_BRP}{(integer)}

\item{B_LU_BBWP}{(numeric) The BBWP category used for allocation of measures to BBWP crop categories}

\item{B_AER_CBS}{(character) The agricultural economic region in the Netherlands (CBS, 2016)}

\item{D_AREA}{(numeric) the area of the field (\ m2 or \ ha)}

\item{B_CT_SOIL}{(numeric) the target value for soil quality conform Ecoregeling scoring}

\item{B_CT_WATER}{(numeric) the target value for water quality conform Ecoregeling scoring}

\item{B_CT_CLIMATE}{(numeric) the target value for climate conform Ecoregeling scoring}

\item{B_CT_BIO}{(numeric) the target value for biodiversity conform Ecoregeling scoring}

\item{B_CT_LANDSCAPE}{(numeric) the target value for landscape quality conform Ecoregeling scoring}
}
\description{
Estimate the actual contribution of crop rotation given aims for soil quality, water quality, climate, biodiversity and landscape
}
