% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbwp_tables.R
\docType{data}
\name{er_crops}
\alias{er_crops}
\title{An overview of crop lists used in Ecoregelingen}
\format{
A data.table with 75 rows and 4 columns:
\describe{
  \item{eco_id}{The unique ID corresponding with a specific Ecoregelingen measure}
  \item{b_lu_brp}{The crop type (conform BRP coding, preferable the most frequent crop on the field)}
  \item{b_lu_name}{The crop name of the most frequent crop on the field}
  \item{er_description}{The description of the crop related measure}
}
}
\usage{
er_crops
}
\description{
This table contains the crop ids (b_lu_brp) for specific Ecoregelingen measures
}
\keyword{datasets}
