% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbwp_tables.R
\docType{data}
\name{er_crops}
\alias{er_crops}
\title{An overview of crop lists used in Ecoregelingen}
\format{
A data.table with 75 rows and 4 columns:
\describe{
  \item{eco_id}{bla bla}
  \item{b_lu_brp}{bla bla}
  \item{b_lu_name}{bla bla}
  \item{er_description}{bla bla}
}
}
\usage{
er_crops
}
\description{
This table contains the crop ids (b_lu_brp) for specific Ecoregelingen measures
}
\keyword{datasets}
