% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbwp_tables.R
\docType{data}
\name{er_farm_measure}
\alias{er_farm_measure}
\title{The impact score of farm based measures}
\format{
A data.table with 40 rows and 5 columns:
\describe{
  \item{id}{A unique ID number}
  \item{indicator}{The environmental theme to which the measure applies}
  \item{eco_id}{The unique ID corresponding with a specific Ecoregelingen measure}
  \item{description}{Short decription of the measure} 
  \item{er_score}{The score of the measure on the environmental theme}
}
}
\usage{
er_farm_measure
}
\description{
This table contains the scores of crop rotation related farm measures for Ecoregeling
}
\keyword{datasets}
