% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbwp_tables.R
\docType{data}
\name{er_farm_measure}
\alias{er_farm_measure}
\title{The impact score of farm based measures}
\format{
A data.table with 40 rows and 5 columns:
\describe{
  \item{id}{bla bla}
  \item{indicator}{bla bla}
  \item{eco_id}{bla bla}
  \item{description}{bla bla}
  \item{er_score}{bla bla}
}
}
\usage{
er_farm_measure
}
\description{
This table contains the scores of crop rotation related farm measures for Ecoregeling
}
\keyword{datasets}
