% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_farm_score.R
\name{er_farm_score}
\alias{er_farm_score}
\title{Calculate the total EcoRegelingen score of five opportunity indicators}
\usage{
er_farm_score(
  S_ER_TOT,
  S_ER_SOIL,
  S_ER_WATER,
  S_ER_CLIMATE,
  S_ER_BIODIVERSITY,
  S_ER_LANDSCAPE,
  S_ER_REWARD,
  B_AREA,
  B_SOILTYPE_AGR,
  sector,
  B_AER_CBS,
  measures
)
}
\arguments{
\item{S_ER_TOT}{(numeric) the Ecoregeling score for the integrative opportunity index for each field}

\item{S_ER_SOIL}{(numeric) the Ecoregeling scoring index for soil quality for each field}

\item{S_ER_WATER}{(numeric) the Ecoregeling scoring index for water quality for each field}

\item{S_ER_CLIMATE}{(numeric) the Ecoregeling scoring index for climate for each field}

\item{S_ER_BIODIVERSITY}{(numeric) the Ecoregeling scoring index for biodiversity for each field}

\item{S_ER_LANDSCAPE}{(numeric) the Ecoregeling scoring index for landscape for each field}

\item{S_ER_REWARD}{(numeric) The financial reward per field for taking Ecoregeling measures (euro / ha)}

\item{B_AREA}{(numeric) the area of the field (m2)}

\item{B_SOILTYPE_AGR}{(character) The type of soil}

\item{sector}{(string) a vector with the farm type given the agricultural sector (options: 'melkveehouderij','akkerbouw','vollegrondsgroente','boomteelt','bollen','veehouderij','overig')}

\item{B_AER_CBS}{(character) The agricultural economic region in the Netherlands (CBS, 2016)}

\item{measures}{(list) the measures planned /done (measure nr)}
}
\description{
Estimate the potential to contribute to agronomic and environmental challenges in a region for a farm
}
