% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_field_scores.R
\name{er_field_scores}
\alias{er_field_scores}
\title{Calculate the total score of five opportunity indicators conform Ecoregelingen Scoring}
\usage{
er_field_scores(
  B_SOILTYPE_AGR,
  B_LU_BBWP,
  B_LU_BRP,
  B_AER_CBS,
  B_AREA,
  B_LU_ARABLE_ER,
  B_LU_PRODUCTIVE_ER,
  B_LU_CULTIVATED_ER,
  B_CT_SOIL,
  B_CT_WATER,
  B_CT_CLIMATE,
  B_CT_BIO,
  B_CT_LANDSCAPE,
  measures = NULL,
  sector
)
}
\arguments{
\item{B_SOILTYPE_AGR}{(character) The type of soil}

\item{B_LU_BBWP}{(character) The BBWP category used for allocation of measures to BBWP crop categories}

\item{B_LU_BRP}{(numeric) The crop code (gewascode) from the BRP}

\item{B_AER_CBS}{(character) The agricultural economic region in the Netherlands (CBS, 2016)}

\item{B_AREA}{(numeric) the area of the field (m2)}

\item{B_LU_ARABLE_ER}{(boolean) does the crop fall within the ER category "arable"}

\item{B_LU_PRODUCTIVE_ER}{(boolean) does the crop fall within the ER category "productive"}

\item{B_LU_CULTIVATED_ER}{(boolean) does the crop fall within the ER category "cultivated"}

\item{B_CT_SOIL}{(numeric) the target value for soil quality conform Ecoregeling scoring}

\item{B_CT_WATER}{(numeric) the target value for water quality conform Ecoregeling scoring}

\item{B_CT_CLIMATE}{(numeric) the target value for climate conform Ecoregeling scoring}

\item{B_CT_BIO}{(numeric) the target value for biodiversity conform Ecoregeling scoring}

\item{B_CT_LANDSCAPE}{(numeric) the target value for landscape quality conform Ecoregeling scoring}

\item{measures}{(list) the measures planned / done per fields (measurement nr)}

\item{sector}{(string) a vector with the farm type given the agricultural sector (options: 'melkveehouderij','akkerbouw','vollegrondsgroente','boomteelt','bollen','veehouderij','overig')}
}
\description{
Estimate the potential to contribute to agronomic and environmental challenges in a region given aims for soil quality, water quality, climate, biodiversity and landscape
}
