% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_meas_rank.R
\name{er_meas_rank}
\alias{er_meas_rank}
\title{Rank the suitability of agronomic measures to improve soil, climate, biodiversity and water management for a given field}
\usage{
er_meas_rank(
  B_SOILTYPE_AGR,
  B_GWL_CLASS,
  A_P_SG,
  B_SLOPE_DEGREE,
  M_DRAIN,
  D_SA_W,
  B_AREA,
  B_AER_CBS,
  B_LU_BBWP,
  B_LU_ECO1,
  B_LU_ECO2,
  B_LU_ECO3,
  B_LU_ECO4,
  B_LU_ECO5,
  B_LU_ECO6,
  B_LU_ECO7,
  B_LU_ECO8,
  B_LU_ECO9,
  B_LU_ECO10,
  B_CT_SOIL,
  B_CT_WATER,
  B_CT_CLIMATE,
  B_CT_BIO,
  B_CT_LANDSCAPE,
  measures,
  sector
)
}
\arguments{
\item{B_SOILTYPE_AGR}{(character) The type of soil}

\item{B_GWL_CLASS}{(character) The groundwater table class}

\item{A_P_SG}{(numeric)}

\item{B_SLOPE_DEGREE}{(numeric) The slope of the field (degrees)}

\item{M_DRAIN}{(boolean) is there tube drainage present in the field}

\item{D_SA_W}{(numeric) The wet perimeter index of the field, fraction that field is surrounded by water}

\item{B_AREA}{(numeric) the area of the field (m2)}

\item{B_AER_CBS}{(character) The agricultural economic region in the Netherlands (CBS, 2016)}

\item{B_LU_BBWP}{(numeric) The BBWP category used for allocation of measures to BBWP crop categories}

\item{B_LU_ECO1}{(boolean) does the crop belong in Ecoregeling category 1}

\item{B_LU_ECO2}{(boolean) does the crop belong in Ecoregeling category 2}

\item{B_LU_ECO3}{(boolean) does the crop belong in Ecoregeling category 3}

\item{B_LU_ECO4}{(boolean) does the crop belong in Ecoregeling category 4}

\item{B_LU_ECO5}{(boolean) does the crop belong in Ecoregeling category 5}

\item{B_LU_ECO6}{(boolean) does the crop belong in Ecoregeling category 6}

\item{B_LU_ECO7}{(boolean) does the crop belong in Ecoregeling category 7}

\item{B_LU_ECO8}{(boolean) does the crop fall within the category "arable"}

\item{B_LU_ECO9}{(boolean) does the crop fall within the category "productive"}

\item{B_LU_ECO10}{(boolean) does the crop fall within the category "cultivated"}

\item{B_CT_SOIL}{(numeric) the target value for soil quality conform Ecoregeling scoring (score / ha)}

\item{B_CT_WATER}{(numeric) the target value for water quality conform Ecoregeling scoring (score / ha)}

\item{B_CT_CLIMATE}{(numeric) the target value for climate conform Ecoregeling scoring (score / ha)}

\item{B_CT_BIO}{(numeric) the target value for biodiversity conform Ecoregeling scoring (score / ha)}

\item{B_CT_LANDSCAPE}{(numeric) the target value for landscape quality conform Ecoregeling scoring (score / ha)}

\item{measures}{(data.table) table with the properties of the available measures}

\item{sector}{(string) a vector with the farm type given the agricultural sector (options: 'diary', 'arable', 'tree_nursery', 'bulbs')}
}
\description{
Estimate the Ecoregelingen score for agronomic measures to improve soil and water management on agricultural farms. 
And send an ordered list back of the most suitable measures.
}
