% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_meas_rank.R
\name{er_meas_rank}
\alias{er_meas_rank}
\title{Rank the suitability of agronomic measures to improve soil, climate, biodiversity and water management for a given field}
\usage{
er_meas_rank(
  B_SOILTYPE_AGR,
  B_LU_BBWP,
  B_GWL_CLASS,
  A_P_SG,
  B_SLOPE,
  B_LU_BRP,
  M_DRAIN,
  D_WP,
  D_AREA,
  B_CT_SOIL,
  B_CT_WATER,
  B_CT_CLIMATE,
  B_CT_BIO,
  B_CT_LANDSCAPE,
  measures,
  sector
)
}
\arguments{
\item{B_SOILTYPE_AGR}{(character) The type of soil}

\item{B_LU_BBWP}{(numeric) The BBWP category used for allocation of measures to BBWP crop categories}

\item{B_GWL_CLASS}{(character) The groundwater table class}

\item{A_P_SG}{(numeric)}

\item{B_SLOPE}{(boolean)}

\item{B_LU_BRP}{(integer)}

\item{M_DRAIN}{(boolean) is there tube drainage present in the field}

\item{D_WP}{(numeric) The fraction of the parcel that is surrounded by surface water}

\item{D_AREA}{(numeric) the area of the field (\ m2 or \ ha)}

\item{B_CT_SOIL}{(numeric) the target value for soil quality conform Ecoregeling scoring}

\item{B_CT_WATER}{(numeric) the target value for water quality conform Ecoregeling scoring}

\item{B_CT_CLIMATE}{(numeric) the target value for climate conform Ecoregeling scoring}

\item{B_CT_BIO}{(numeric) the target value for biodiversity conform Ecoregeling scoring}

\item{B_CT_LANDSCAPE}{(numeric) the target value for landscape quality conform Ecoregeling scoring}

\item{measures}{(data.table) table with the properties of the available measures}

\item{sector}{(string) a vector with the farm type given the agricultural sector (options: 'diary', 'arable', 'tree_nursery', 'bulbs')}
}
\description{
Estimate the Ecoregelingen score for agronomic measures to improve soil and water management on agricultural farms. 
And send an ordered list back of the most suitable measures.
}
