% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_meas_scores.R
\name{er_meas_score}
\alias{er_meas_score}
\title{Evaluate the contribution of agronomic measures to improve soil mand water management}
\usage{
er_meas_score(B_SOILTYPE_AGR, B_LU_BRP, B_LU_BBWP, measures, sector)
}
\arguments{
\item{B_SOILTYPE_AGR}{(character) The type of soil}

\item{B_LU_BRP}{(numeric) The crop type (conform BRP coding, preferable the most frequent crop on the field)}

\item{B_LU_BBWP}{(numeric) The BBWP category used for allocation of measures to BBWP crop categories}

\item{measures}{(list) the measures planned / done per fields}

\item{sector}{(string) a vector with the farm type given the agricultural sector (options: 'diary', 'arable', 'tree_nursery', 'bulbs')}
}
\description{
Estimate the Ecoregeling score for agronomic measures taken to improve soil and water management on agricultural farms.
}
