% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbwp_tables.R
\docType{data}
\name{er_scoring}
\alias{er_scoring}
\title{The importance and scoring of environmental challenges for Ecoregelingen}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 8 rows and 7 columns.
}
\usage{
er_scoring
}
\description{
This table contains the objective scoring and correction factors to estimate score for Ecoregeling
}
\details{
\describe{
  \item{soiltype}{Agricultural soil type}
  \item{type}{The type of weighting factor: urgency (distribution of scores depending on political urgency) or aim (distribution of scores depending on aim for environmental challenges)}
  \item{cf_soil}{Weighting factor for soil}
  \item{cf_water}{Weighting factor for water}
  \item{cf_climate}{Weighting factor for climate}
  \item{cf_biodiversity}{Weighting factor for biodiversity}
  \item{cf_landscape}{Weighting factor for landscape}
}
}
\keyword{datasets}
