% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbwp_tables.R
\docType{data}
\name{er_scoring}
\alias{er_scoring}
\title{The importance and scoring of environmental challenges for Ecoregelingen}
\format{
A data.table with 8 rows and 7 columns:
\describe{
  \item{soiltype}{The type of soil}
  \item{type}{The type of driving factor (either political urgency (urgency) or contribution to an environmental theme (aim))}
  \item{cf_soil}{Correction factor to estimate soil score for Ecoregeling}
  \item{cf_water}{Correction factor to estimate water score for Ecoregeling}
  \item{cf_climate}{Correction factor to estimate climate score for Ecoregeling}
  \item{cf_biodiversity}{Correction factor to estimate biodiversity score for Ecoregeling}
  \item{cf_landscape}{Correction factor to estimate landscape score for Ecoregeling}
}
}
\usage{
er_scoring
}
\description{
This table contains the objective scoring and correction factors to estimate score for Ecoregeling
}
\keyword{datasets}
